+++ src/components/Email/Email.module.scss
@@ -1,0 +1,5 @@
+.error {
+  font-size: 0.7rem;
+  color: red;
+  margin-top: -0.5rem;
+}
