+++ src/components/auth/RegistrationBenefits/RegistrationBenefits.module.scss
