+++ src/pages/SignIn/index.tsx
@@ -1,8 +1,8 @@
 import { AuthHeader } from '../../components/auth/AuthHeader';
-import { AuthForm } from '../../components/auth/AuthForm';
 import { AuthButton } from '../../components/auth/AuthButton';
-import { AuthLink } from '../../components/auth/AuthLinks'; 
-import { RegistrationBenefits } from '../../components/RegistrationBenefits';
+import { AuthLink } from '../../components/auth/AuthLinks';
+import { LoginForm } from '../../components/auth/AuthForm/LoginForm';
+import { RegistrationBenefits } from '../../components/auth/RegistrationBenefits';
 import s from './SignIn.module.scss';
 
 export const SignIn = () => {
@@ -11,12 +11,8 @@ export const SignIn = () => {
       <RegistrationBenefits />
       <div className={s.wrap}>
         <AuthHeader text='Вхід' />
-        <AuthForm />
-        <AuthButton
-          text='Sign in with Google'
-          className={s.btn}
-          variant='secondary'
-        />
+        <LoginForm />
+        <AuthButton text='Sign in with Google' className={s.btn} variant='secondary' />
         <AuthLink />
       </div>
     </section>
